requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
streamlit
langchain_google_genai
